# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Enrichment API",
    "description": "API for file enrichment services",
    "version": "0.1.0"
  },
  "paths": {
...
    "/dpapi/credentials": {
      "post": {
        "tags": [
          "dpapi"
        ],
        "summary": "Submit DPAPI credential for masterkey decryption",
        "description": "Submit credential material to decrypt DPAPI master keys. Supports passwords, NTLM hashes, cred keys, domain backup keys, decrypted master keys, and Chromium app-bound-encryption keys.",
        "operationId": "submit_dpapi_credential_dpapi_credentials_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DpapiCredentialRequest",
                "description": "The DPAPI credential data"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChromiumAppBoundKeyCredential": {
        "properties": {
          "type": {
            "type": "string",
            "const": "chromium_app_bound_key",
            "title": "Type"
          },
          "value": {
            "type": "string",
            "title": "Value"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "browser": {
            "type": "string",
            "enum": [
              "chrome",
              "edge",
              "brave",
              "opera"
            ],
            "title": "Browser",
            "default": "chrome"
          },
          "username": {
            "type": "string",
            "title": "Username",
            "default": "UNKNOWN"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "type",
          "value",
          "source"
        ],
        "title": "ChromiumAppBoundKeyCredential",
        "description": "Chromium App-Bound-Encryption key credential."
      },
...
      "DomainBackupKeyCredential": {
        "properties": {
          "type": {
            "type": "string",
            "const": "domain_backup_key",
            "title": "Type"
          },
          "value": {
            "type": "string",
            "title": "Value"
          },
          "guid": {
            "type": "string",
            "title": "Guid"
          },
          "domain_controller": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Domain Controller"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "type",
          "value",
          "guid"
        ],
        "title": "DomainBackupKeyCredential",
        "description": "Domain backup key credential (PVK format)."
      },
      "DpapiCredentialRequest": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/PasswordCredentialKey"
          },
          {
            "$ref": "#/components/schemas/NtlmHashCredentialKey"
          },
          {
            "$ref": "#/components/schemas/Sha1CredentialKey"
          },
          {
            "$ref": "#/components/schemas/Pbkdf2StrongCredentialKey"
          },
          {
            "$ref": "#/components/schemas/DomainBackupKeyCredential"
          },
          {
            "$ref": "#/components/schemas/MasterKeyGuidPairList"
          },
          {
            "$ref": "#/components/schemas/DpapiSystemCredentialRequest"
          },
          {
            "$ref": "#/components/schemas/ChromiumAppBoundKeyCredential"
          }
        ]
      },
      "DpapiSystemCredentialRequest": {
        "properties": {
          "type": {
            "type": "string",
            "const": "dpapi_system",
            "title": "Type"
          },
          "value": {
            "type": "string",
            "pattern": "^[0-9a-fA-F]{80}$",
            "title": "Value",
            "description": "Hex-encoded DPAPI_SYSTEM LSA secret (40 bytes)"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "type",
          "value"
        ],
        "title": "DpapiSystemCredentialRequest",
        "description": "DPAPI_SYSTEM LSA Secret credential sent in an API request."
      },
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "Error message details"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MasterKeyGuidPair": {
        "properties": {
          "guid": {
            "type": "string",
            "format": "uuid",
            "title": "Guid",
            "description": "Master key GUID"
          },
          "key_hex": {
            "type": "string",
            "pattern": "^[0-9a-fA-F]+$",
            "title": "Key Hex",
            "description": "Hex-encoded master key bytes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "guid",
          "key_hex"
        ],
        "title": "MasterKeyGuidPair",
        "description": "Strongly typed master key data."
      },
      "MasterKeyGuidPairList": {
        "properties": {
          "type": {
            "type": "string",
            "const": "master_key_guid_pair",
            "title": "Type"
          },
          "value": {
            "items": {
              "$ref": "#/components/schemas/MasterKeyGuidPair"
            },
            "type": "array",
            "title": "Value"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "type",
          "value"
        ],
        "title": "MasterKeyGuidPairList",
        "description": "Decrypted master key/GUID pairs."
      },
      "NtlmHashCredentialKey": {
        "properties": {
          "type": {
            "type": "string",
            "const": "cred_key_ntlm",
            "title": "Type"
          },
          "value": {
            "type": "string",
            "title": "Value"
          },
          "user_sid": {
            "type": "string",
            "title": "User Sid"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "type",
          "value",
          "user_sid"
        ],
        "title": "NtlmHashCredentialKey",
        "description": "Use an NTLM hash to derive a DPAPI credential key."
      },
      "PasswordCredentialKey": {
        "properties": {
          "type": {
            "type": "string",
            "const": "password",
            "title": "Type"
          },
          "value": {
            "type": "string",
            "title": "Value"
          },
          "user_sid": {
            "type": "string",
            "title": "User Sid"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "type",
          "value",
          "user_sid"
        ],
        "title": "PasswordCredentialKey",
        "description": "Use a plaintext password to derive a DPAPI credential key."
      },
      "Pbkdf2StrongCredentialKey": {
        "properties": {
          "type": {
            "type": "string",
            "const": "cred_key_pbkdf2",
            "title": "Type"
          },
          "value": {
            "type": "string",
            "title": "Value"
          },
          "user_sid": {
            "type": "string",
            "title": "User Sid"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "type",
          "value",
          "user_sid"
        ],
        "title": "Pbkdf2StrongCredentialKey",
        "description": "Credential key object."
      },
...
      "Sha1CredentialKey": {
        "properties": {
          "type": {
            "type": "string",
            "const": "cred_key_sha1",
            "title": "Type"
          },
          "value": {
            "type": "string",
            "title": "Value"
          },
          "user_sid": {
            "type": "string",
            "title": "User Sid"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "type",
          "value",
          "user_sid"
        ],
        "title": "Sha1CredentialKey",
        "description": "Use a SHA1 hash to derive a DPAPI credential key."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  },
...
}